backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=1.10.0,<2.0.0

# AI and ML
google-generativeai>=0.8.0

# PDF Processing
pdf2image>=1.16.0
fillpdf>=0.7.0
pillow>=10.0.0
PyMuPDF>=1.23.0

# Environment and utilities
python-dotenv>=1.0.0

# Additional dependencies
requests>=2.28.0